requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
django-cors-headers==4.9.0
django-nested-admin==4.1.6
django-tinymce==5.0.0
django_dashub==4.0.0
django_filehub==3.2.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-decouple==3.8
python-dotenv==1.2.1
python-monkey-business==1.1.0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sqlparse==0.5.5
standard-imghdr==3.13.0
uritemplate==4.2.0
urllib3==2.6.3
